requirements.txt
tweepy
Pillow
requests